moved container to separate module, updated footer, pulled in sidebar changes
