+++ src/app/public/src/modules/container/container.component.scss
