+++ src/app/public/src/modules/container/container.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { StacheContainerComponent } from './container.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [
+    StacheContainerComponent
+  ],
+  exports: [
+    StacheContainerComponent
+  ],
+  providers: []
+})
+export class StacheContainerModule { }
