+++ src/app/public/src/modules/container/index.ts
@@ -1,0 +1,2 @@
+export * from './container.component';
+export * from './container.module';
