+++ src/app/public/src/modules/footer/footer.module.ts
@@ -2,7 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { StacheFooterComponent } from './footer.component';
+
 import { StacheLinkModule } from '../link';
+import { StacheContainerModule } from '../container';
 
 @NgModule({
   declarations: [
@@ -10,7 +12,8 @@ import { StacheLinkModule } from '../link';
   ],
   imports: [
     CommonModule,
-    StacheLinkModule
+    StacheLinkModule,
+    StacheContainerModule
   ],
   exports: [
     StacheFooterComponent
