+++ src/app/public/src/modules/layout/index.ts
@@ -1,3 +1,2 @@
-export * from './container.component';
 export * from './layout.component';
 export * from './layout.module';
